requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
emoji==2.2.0
Flask==2.3.2
gh-md-to-html==1.21.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
pip==23.0.1
pymongo==4.3.3
pyrankvote==2.0.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
setuptools==66.0.0
shellescape==3.8.1
soupsieve==2.4.1
tabulate==0.9.0
urllib3==2.0.2
webcolors==1.13
Werkzeug==2.3.3
wheel==0.38.4
